remove fluid.ir.RegisterPassHelper PassDesc and RegisterPass
